modbus-simulator/requirements.txt
pymodbus>=3.6.0
pyyaml>=6.0
